[#173][coupon] refactor: 스프링 이벤트 리스너 인프라스트럭쳐로 이동
